Filter links using full resource links
Instead of using relative links from a page,
we filter the links using the full link to
prevent not downloading and also downloading
multiple files.
For instance ../hello.js from page hello.html
can point to the same file as /hello.js in
home.html with hello.html being a sub-page of
home.html(home/hello).
The two relative links point to the same file
but filtering using the relative link won't
filter that case out.

Also, the replacement of links is done using
a combination of the element attribute and the
relative link. For example href="/hello.js"
instead of just /hello.js.
This is to prevent the case where replacing a
relative link such as "/" would replace all /
in the html document